requirements.txt
# Core Deep Learning Framework
torch==2.1.0
torchvision==0.16.0

# Data Processing
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
opencv-python==4.8.0.74

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Model Evaluation
scikit-learn==1.3.0

# Progress Bars
tqdm==4.66.1

# Jupyter Notebook
jupyter==1.0.0
ipywidgets==8.1.0

# Utilities
PyYAML==6.0.1

# Kaggle API for dataset download
kaggle==1.6.6
